+++ src/main/scala/engine/core/Node.scala
@@ -1,9 +1,10 @@
-package engine
+package engine.core
 
 import scala.reflect.ClassTag
 import engine.math.Vector2
 import engine.Component
 import lib.instance_management.Instance
+import core.Game
 
 class Node private[engine] (game: Game) {
