+++ src/main/scala/engine/scene/Scene.scala
@@ -1,5 +1,5 @@
 package engine.scene
 
-import engine.Game
+import engine.core.Game
 
 trait Scene(game: Game) extends EngineListener
